test(ui): hoist shared worker-middleware harness + passthrough suite (Sonar dup)
The prior stubWorker extraction only trimmed duplication to 12.1% (gate is 3%):
the two middleware tests still shared an identical setup() body and the two
"forwards non-worker actions" test cases. Move both into src/test/workerStub.ts
as createWorkerHarness() (stubs Worker + builds the dispatch/next/api spies) and
itForwardsNonWorkerActions(setup) (the shared passthrough it-blocks). Each test
now keeps only its middleware-specific setup line and its own cases; behavior
and 100% coverage of both middlewares are unchanged.
