+++ ui/src/store/features/previewsWorker/previewsWorkerMiddleware.test.ts
@@ -13,8 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import type { MiddlewareAPI } from '@reduxjs/toolkit';
+import { afterEach, describe, expect, it, vi } from 'vitest';
 import { previewsWorkerMiddleware } from './previewsWorkerMiddleware.ts';
 import {
   getReactionActions,
@@ -23,14 +22,10 @@ import {
 } from 'store/entities/reactions/reactions.actions.ts';
 import { getAllTemplatesActions } from 'store/entities/templates/templates.actions.ts';
 import { setPreviewsByIds } from 'store/entities/reactions/reactionsPreviews/reactionsPreviews.actions.ts';
-import { type CapturedWorker, stubWorker } from 'test/workerStub.ts';
-
-let workers: Array<CapturedWorker>;
+import { createWorkerHarness, itForwardsNonWorkerActions } from 'test/workerStub.ts';
 
 function setup() {
-  const dispatch = vi.fn((action: unknown) => action);
-  const api = { dispatch, getState: vi.fn() } as unknown as MiddlewareAPI;
-  const next = vi.fn((action: unknown) => action);
+  const { workers, dispatch, next, api } = createWorkerHarness();
   const invoke = previewsWorkerMiddleware(api)(next) as (action: unknown) => unknown;
   return { dispatch, next, invoke, worker: workers[0] };
 }
@@ -40,22 +35,12 @@ function setup() {
 // without fabricating full DatasetReaction/Pages payloads.
 const action = (type: string, payload: unknown) => ({ type, payload });
 
-beforeEach(() => {
-  workers = stubWorker();
-});
-
 afterEach(() => {
   vi.unstubAllGlobals();
 });
 
 describe('previewsWorkerMiddleware', () => {
-  it('passes thunk (function) actions straight through without posting to the worker', () => {
-    const { invoke, next, worker } = setup();
-    const thunk = () => undefined;
-    invoke(thunk);
-    expect(next).toHaveBeenCalledWith(thunk);
-    expect(worker.postMessage).not.toHaveBeenCalled();
-  });
+  itForwardsNonWorkerActions(setup);
 
   it("posts a single reaction action's previews to the worker and forwards the action", () => {
     const { invoke, next, worker } = setup();
@@ -84,14 +69,6 @@ describe('previewsWorkerMiddleware', () => {
     const { invoke, worker } = setup();
     invoke(action(getAllTemplatesActions.success.type, [{ previews: { t1: 'x' } }, { previews: { t2: 'y' } }]));
     expect(worker.postMessage).toHaveBeenCalledWith({ t1: 'x', t2: 'y' });
-  });
-
-  it('does not post unrelated actions but still forwards them', () => {
-    const { invoke, next, worker } = setup();
-    const unrelated = action('something/unrelated', undefined);
-    invoke(unrelated);
-    expect(worker.postMessage).not.toHaveBeenCalled();
-    expect(next).toHaveBeenCalledWith(unrelated);
   });
 
   it('dispatches setPreviewsByIds when the worker posts a result back', () => {
